🔥: Moving ProjectSearch up into its own file
